+++ R/GermValue.R
@@ -44,7 +44,8 @@
 #' Emergence energy \insertCite{bonner_ideal_1967}{germinationmetrics}.
 #'
 #' For daily germination counts, germination value (\mjseqn{GV}) is computed as
-#' follows \insertCite{czabator_germination_1962}{germinationmetrics}.
+#' follows \insertCite{czabator_germination_1962}{germinationmetrics} (Specified
+#' by the argument \code{method = "czabator"}).
 #'
 #' \mjsdeqn{GV = PV \times MDG}
 #'
@@ -57,7 +58,7 @@
 #'
 #' A new estimation of germination value was given by
 #' \insertCite{djavanshir_germination_1976;textual}{germinationmetrics} as
-#' follows.
+#' follows (Specified by the argument \code{method = "dp"}).
 #'
 #' \mjsdeqn{GV = \frac{\sum DGS}{N} \times GP \times c}
 #'