+++ NEWS.md
@@ -6,6 +6,7 @@
 ## UPDATED FUNCTIONS:
 * `FirstGermTime`, `LastGermTime` and `PeakGermTime` - Updated documentation with formula.
 * `EmergenceRateIndex` - Depreciate erroneous `method` argument options of `melville` and `melvillesantanaranal` and replace them with `shmueligoldberg` and `sgsantanaranal` respectively.
+ `GermSpeedCorrected` - Use germination percentage instead of counts.
 
 ## VIGNETTE:
 * Fixed fetching CRAN page for version history to fail gracefully when access is down.