+++ NEWS.md
@@ -1,3 +1,5 @@
+# germinationmetrics  0.1.7
+
 # germinationmetrics  0.1.6
 
 ## NEW FUNCTIONS: